Sacrilege Beta 0.1.0!!!!
Added the Riftcaller
Added Cults
Finished up the Blood Bottle
Added Sculk Covered Rocks, Items, and Armor to the Ancient City loot table

+++ src/main/java/com/ozymandiu5/sacrilege/items/AltarItem.java
@@ -49,7 +49,7 @@ public class AltarItem extends Item {
 		}
 
 		public static void setAltarType(ItemStack stack, int altarType) {
-			stack.getOrCreateTag().putFloat(TAG_ALTARTYPE, altarType);
+			stack.getOrCreateTag().putInt(TAG_ALTARTYPE, altarType);
 		}
 		
 	   public InteractionResult useOn(UseOnContext p_40510_) {
